moved styles to commons project unified ADVColor, ADVStrokeStyle, ADVStrokeThickness, ADVStyle, ADVValueStyle, ADVDefaultStyle, ADVDefaultLineStyle, ADVException, ADVStyleException used default styles from lib (potentially contains nullpointer, because of unified default initialization) => test seems good though
