+++ lib-core/src/main/java/ch/hsr/adv/lib/core/logic/domain/styles/presets/ADVErrorStyle.java
@@ -1,13 +1,14 @@
 package ch.hsr.adv.lib.core.logic.domain.styles.presets;
 
-import ch.hsr.adv.lib.core.logic.domain.styles.ADVStrokeStyle;
-import ch.hsr.adv.lib.core.logic.domain.styles.ADVStrokeThickness;
-import ch.hsr.adv.lib.core.logic.domain.styles.ADVStyle;
+import ch.hsr.adv.commons.core.logic.domain.styles.ADVStrokeStyle;
+import ch.hsr.adv.commons.core.logic.domain.styles.ADVStrokeThickness;
+import ch.hsr.adv.commons.core.logic.domain.styles.ADVStyle;
 
 /**
  * A predefined style class with a fat red border.
  */
 public class ADVErrorStyle implements ADVStyle {
+
     private static final transient int LIGHT_RED = 0xd5b9bb;
     private static final transient int DARK_RED = 0xb95f6a;
     private final int fillColor;
