--- lib-core/src/main/java/ch/hsr/adv/lib/core/logic/util/ADVStyleException.java
@@ -1,11 +1,0 @@
-package ch.hsr.adv.lib.core.logic.util;
-
-/**
- * To be thrown for any exception to do with the style of an element.
- */
-public class ADVStyleException extends ADVException {
-
-    public ADVStyleException(String message) {
-        super(message);
-    }
-}
