+++ module-bb-stack/src/main/java/ch/hsr/adv/lib/stack/logic/StackModule.java
@@ -1,8 +1,8 @@
 package ch.hsr.adv.lib.stack.logic;
 
+import ch.hsr.adv.commons.core.logic.domain.styles.ADVStyle;
 import ch.hsr.adv.commons.stack.logic.ConstantsStack;
 import ch.hsr.adv.lib.core.logic.ADVModule;
-import ch.hsr.adv.lib.core.logic.domain.styles.ADVStyle;
 import ch.hsr.adv.lib.core.logic.util.StringUtil;
 import ch.hsr.adv.lib.stack.logic.domain.ADVStack;
