+++ module-bb-stack/src/main/java/ch/hsr/adv/lib/stack/logic/domain/StackElement.java
@@ -1,8 +1,8 @@
 package ch.hsr.adv.lib.stack.logic.domain;
 
 
+import ch.hsr.adv.commons.core.logic.domain.styles.ADVStyle;
 import ch.hsr.adv.lib.core.logic.domain.ADVElement;
-import ch.hsr.adv.lib.core.logic.domain.styles.ADVStyle;
 
 /**
  * Represents the content of one stack element.
